requirements.txt
asgiref==3.5.2
astroid==2.11.5
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.5.1
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-heroku==0.3.1
django-on-heroku==1.1.2
djangorestframework==3.13.1
drf-yasg==1.20.0
Faker==13.12.0
faker-commerce==1.0.3
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==21.3
Pillow==9.1.1
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pylint==2.14.0
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
wrapt==1.14.1
